% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_grob}
\alias{add_grob}
\title{Add a grob to a gtable}
\usage{
add_grob(
  gtable,
  grob,
  position = c("top", "bottom", "left", "right", "none"),
  space = NULL,
  clip = "on"
)
}
\arguments{
\item{gtable}{A gtable object.}

\item{grob}{A grob or gtable object to add.}

\item{position}{The position to add the grob.
One of \code{"top"}, \code{"bottom"}, \code{"left"}, \code{"right"}, or \code{"none"}.}

\item{space}{The space to allocate for the grob.
If \code{NULL}, will be calculated automatically.}

\item{clip}{The clipping mode.
Default is \code{"on"}.}
}
\value{
A gtable object with the grob added.
}
\description{
Add a grob to a gtable at a specified position (top, bottom, left, or right).
}
