% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{annotation_graphics}
\alias{annotation_graphics}
\title{Build graphics callback list for anno_customize}
\usage{
annotation_graphics(subplots, prefix)
}
\arguments{
\item{subplots}{A named list of subplot objects.}

\item{prefix}{Prefix used to generate stable grob names.}
}
\value{
A named list of callback functions accepted by \link[ComplexHeatmap:anno_customize]{ComplexHeatmap::anno_customize}.
}
\description{
Build graphics callback list for anno_customize
}
