% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_grob}
\alias{as_grob}
\title{Convert a plot object to a grob}
\usage{
as_grob(plot, ...)
}
\arguments{
\item{plot}{A plot object (gList, patchwork, or ggplot).}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A grob object.
}
\description{
Convert various plot objects (gList, patchwork, ggplot) to a grob object.
}
