% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_gtable}
\alias{as_gtable}
\title{Convert a plot object to a gtable}
\usage{
as_gtable(plot, ...)
}
\arguments{
\item{plot}{A plot object (gtable, grob, patchwork, or ggplot).}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A gtable object.
}
\description{
Convert various plot objects (gtable, grob, patchwork, ggplot) to a gtable object.
}
