% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{Blend2Color}
\alias{Blend2Color}
\title{Blend two colors using a specified mode}
\usage{
Blend2Color(C1, C2, mode = "blend")
}
\arguments{
\item{C1}{A list containing the first color RGB values and alpha channel.}

\item{C2}{A list containing the second color RGB values and alpha channel.}

\item{mode}{The blending mode to use.
One of \code{"blend"}, \code{"average"}, \code{"screen"}, or \code{"multiply"}.
Default is \code{"blend"}.}
}
\value{
A list containing the blended RGB values and alpha channel.
}
\description{
Blend two colors with alpha channels using one of several blending modes:
blend, average, screen, or multiply.
}
