% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{BlendRGBList}
\alias{BlendRGBList}
\title{Blend a list of colors}
\usage{
BlendRGBList(Clist, mode = "blend", RGB_BackGround = c(1, 1, 1))
}
\arguments{
\item{Clist}{A list of colors, where each color is a list containing
RGB values and alpha channel.}

\item{mode}{The blending mode to use.
One of \code{"blend"}, \code{"average"}, \code{"screen"}, or \code{"multiply"}.
Default is \code{"blend"}.}

\item{RGB_BackGround}{The background RGB color to composite with.
Default is \code{c(1, 1, 1)} (white).}
}
\value{
A numeric vector of RGB values.
}
\description{
Blend multiple colors with alpha channels into a single color using
a specified blending mode.
}
