% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{col2hex}
\alias{col2hex}
\title{Convert color names to hexadecimal format}
\usage{
col2hex(cname)
}
\arguments{
\item{cname}{A character vector of color names.}
}
\value{
A character vector of hexadecimal color codes.
}
\description{
Convert color names to hexadecimal RGB color codes.
}
