% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{extractgrobs}
\alias{extractgrobs}
\title{Extract grobs from a list}
\usage{
extractgrobs(vlnplots, x_nm, y_nm, x, y)
}
\arguments{
\item{vlnplots}{A named list of grobs.}

\item{x_nm}{A character vector of names for the x dimension.}

\item{y_nm}{A character vector of names for the y dimension.}

\item{x}{An integer index for the x dimension.}

\item{y}{An integer index for the y dimension.}
}
\value{
The extracted grob(s).
}
\description{
Extract grobs from a named list of grobs based on the specified x and y indices.
}
