% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{geom_alluvial}
\alias{geom_alluvial}
\title{geom_alluvial}
\usage{
geom_alluvial(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  space = 0,
  width = 0.1,
  smooth = 8,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Provide you own mapping. Both x and y need to be numeric.}

\item{data}{Provide you own data.}

\item{position}{Change position.}

\item{na.rm}{Remove missing values.}

\item{show.legend}{Show legend in plot.}

\item{space}{Space between nodes in the y-direction.}

\item{width}{Width of nodes.}

\item{smooth}{How much smooth should the curve have? More means steeper curve.}

\item{inherit.aes}{Should the geom inherit aesthetics.}

\item{...}{Other arguments to be passed to the geom.}
}
\value{
A ggplot layer.
}
\description{
Creates an alluvial plot which visualize flows between nodes.
Each observation needs to have a `x` aesthetic as well as a `next_x` column which declares where that observation should flow.
Also each observation should have a `node` and a `next_node` aesthetic which provide information about which group in the y-direction.
}
