% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{geom_alluvial_label}
\alias{geom_alluvial_label}
\alias{geom_alluvial_text}
\title{geom_alluvial_label}
\usage{
geom_alluvial_text(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  space = 0,
  width = 0.1,
  inherit.aes = TRUE,
  ...
)

geom_alluvial_label(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  space = 0,
  width = 0.1,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Provide you own mapping. Both x and y need to be numeric.}

\item{data}{Provide you own data.}

\item{position}{Change position.}

\item{na.rm}{Remove missing values.}

\item{show.legend}{Show legend in plot.}

\item{space}{Space between nodes in the y-direction.}

\item{width}{Width of nodes.}

\item{inherit.aes}{Should the geom inherit aesthetics.}

\item{...}{Other arguments to be passed to the geom.}
}
\value{
A ggplot layer.
}
\description{
Creates centered labels or text in nodes of your alluvial plot.
Needs to have the exact same aesthetics as the call to `geom_alluvial` to work.
}
\details{
Other important arguments are;
`space` which provides the space between nodes in the y-direction;
`shift` which shifts nodes in the y-direction.
}
