% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{geom_sankey_bump}
\alias{geom_sankey_bump}
\title{geom_sankey_bump}
\usage{
geom_sankey_bump(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  smooth = 8,
  type = "sankey",
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Provide you own mapping. Both x and y need to be numeric.}

\item{data}{Provide you own data.}

\item{position}{Change position.}

\item{na.rm}{Remove missing values.}

\item{show.legend}{Show legend in plot.}

\item{smooth}{How much smooth should the curve have? More means steeper curve.}

\item{type}{Either `sankey` or `alluvial`.}

\item{inherit.aes}{Should the geom inherit aesthetics.}

\item{...}{Other arguments to be passed to the geom.}
}
\value{
A ggplot layer.
}
\description{
Creates a sankey bump plot which visualize flows between nodes.
Each observation needs to have a `x` aesthetic as well as a `next_x` column which declares where that observation should flow.
Also, each observation should have a `node` and a `next_node` aesthetic which provide information about which group in the y-direction.
}
\details{
Other important arguments are;
`space` which provides the space between nodes in the y-direction;
`shift` which shifts nodes in the y-direction.
}
