% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_chinese_palettes}
\alias{get_chinese_palettes}
\title{Get Chinese color palettes}
\usage{
get_chinese_palettes(prefix = "Chinese")
}
\arguments{
\item{prefix}{The prefix of the palette names.
Default is `"Chinese_"`.}
}
\value{
A list of Chinese color palettes.
}
\description{
Get Chinese color palettes
}
\examples{
show_palettes(get_chinese_palettes())
}
