% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{grid_draw}
\alias{grid_draw}
\title{Draw grobs at specified positions}
\usage{
grid_draw(groblist, x, y, width, height)
}
\arguments{
\item{groblist}{A grob or a list of grobs to draw.}

\item{x}{A numeric vector of x positions for each grob.}

\item{y}{A numeric vector of y positions for each grob.}

\item{width}{A numeric vector of widths for each grob.}

\item{height}{A numeric vector of heights for each grob.}
}
\description{
Draw a list of grobs at specified positions with given widths and heights.
}
