% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mestimate}
\alias{mestimate}
\title{Estimate the fuzzifier parameter m}
\usage{
mestimate(data)
}
\arguments{
\item{data}{A matrix or data frame.}
}
\value{
The estimated fuzzifier parameter m.
}
\description{
Estimate the fuzzifier parameter m for fuzzy clustering based on the data dimensions.
}
