% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{no_legend}
\alias{no_legend}
\title{Remove the legend}
\usage{
no_legend()
}
\value{
A ggplot2 theme object.
}
\description{
Remove the legend
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
  geom_point() +
  no_legend()
}
