% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{no_margin}
\alias{no_margin}
\title{Remove plot margins}
\usage{
no_margin()
}
\value{
A ggplot2 theme object.
}
\description{
Remove plot margins
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  no_margin()
}
