% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{no_x_text}
\alias{no_x_text}
\title{Remove x-axis text and ticks}
\usage{
no_x_text()
}
\value{
A ggplot2 theme object.
}
\description{
Remove x-axis text and ticks
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  no_x_text()
}
