% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{no_y_text}
\alias{no_y_text}
\title{Remove y-axis text and ticks}
\usage{
no_y_text()
}
\value{
A ggplot2 theme object.
}
\description{
Remove y-axis text and ticks
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  no_y_text()
}
