% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{normalize_drawable}
\alias{normalize_drawable}
\title{Normalize drawable objects to grobs}
\usage{
normalize_drawable(obj)
}
\arguments{
\item{obj}{A drawable object, such as a ggplot, patchwork, or grid grob.}
}
\value{
A grid grob object. If the input is not drawable, returns a null grob.
}
\description{
Normalize drawable objects to grobs
}
\examples{
library(ggplot2)
p <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
  ggplot2::geom_point()
g <- normalize_drawable(p)
p + g
}
