% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_to_sum}
\alias{normalize_to_sum}
\title{Normalize values to a unit sum}
\usage{
normalize_to_sum(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector scaled to sum to 1.
}
\description{
Normalize values to a unit sum
}
\keyword{internal}
