% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{patchwork_grob}
\alias{patchwork_grob}
\title{Convert a patchwork object to a grob}
\usage{
patchwork_grob(x, ...)
}
\arguments{
\item{x}{A patchwork object.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A gtable object.
}
\description{
Convert a patchwork object to a gtable grob by processing annotations
and building the patchwork layout.
}
