% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese_colors.R
\name{print.ChineseColorInfo}
\alias{print.ChineseColorInfo}
\title{Print ChineseColorInfo object}
\usage{
\method{print}{ChineseColorInfo}(x, ...)
}
\arguments{
\item{x}{A ChineseColorInfo object.}

\item{...}{Additional arguments passed to print.}
}
\value{
Details of the ChineseColorInfo object.
}
\description{
Print ChineseColorInfo object
}
