% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese_colors.R
\name{print.ChineseColors}
\alias{print.ChineseColors}
\title{Print \code{ChineseColors} object}
\usage{
\method{print}{ChineseColors}(x, ...)
}
\arguments{
\item{x}{A \code{ChineseColors} object.}

\item{...}{Additional arguments.}
}
\value{
Details of the \code{ChineseColors} object.
}
\description{
Print \code{ChineseColors} object
}
