% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{print.colors}
\alias{print.colors}
\title{Print colors object}
\usage{
\method{print}{colors}(x, ...)
}
\arguments{
\item{x}{A colors object (data frame with color information).}

\item{...}{Additional arguments passed to print.}
}
\value{
Details of the colors objec.
}
\description{
Print colors object
}
