% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_this}
\alias{theme_this}
\title{The default theme for scop plot function.}
\usage{
theme_this(aspect.ratio = NULL, base_size = 12, ...)
}
\arguments{
\item{aspect.ratio}{Aspect ratio of the panel.}

\item{base_size}{Base font size}

\item{...}{Arguments passed to the \link[ggplot2:theme]{ggplot2::theme}.}
}
\description{
The default theme for scop plot function.
}
\examples{
library(ggplot2)
p <- ggplot(
  data = mtcars,
  aes(x = wt, y = mpg, colour = factor(cyl))
) +
  geom_point()
p + theme_this()
}
