% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisplot-package.R
\docType{package}
\name{thisplot-package}
\alias{thisplot}
\alias{thisplot-package}
\title{Utility Functions for Data Visualization and Plotting}
\source{
\url{https://mengxu98.github.io/thisplot/}
}
\description{
Provides utility functions for data visualization and plotting in R. Includes functions for color manipulation, plot customization, panel size control, data optimization for plots, and layout adjustments. Designed to enhance workflows with ggplot2, patchwork, and ComplexHeatmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/thisplot/}
  \item Report bugs at \url{https://github.com/mengxu98/thisplot/issues}
}

}
\author{
Meng Xu (Maintainer), \email{mengxu98@qq.com}
}
