% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_percent}
\alias{to_percent}
\title{Convert proportions to percentages}
\usage{
to_percent(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector expressed as percentages.
}
\description{
Convert proportions to percentages
}
\examples{
to_percent(c(1, 1, 2))
to_percent(c(0, 0))
}
