% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{annotation_block_fill_graphics}
\alias{annotation_block_fill_graphics}
\title{Build block fill panel function}
\usage{
annotation_block_fill_graphics(
  levels,
  palette = NULL,
  palcolor = NULL,
  fill_values = NULL,
  border = FALSE
)
}
\arguments{
\item{levels}{Character/factor levels used as keys for fill mapping.}

\item{palette}{Palette name.
All available palette names can be queried with \link{show_palettes}.}

\item{palcolor}{Custom colors used to create a color palette.}

\item{fill_values}{Optional named fill vector. Names should match \code{levels}.}

\item{border}{Whether to draw block border.}
}
\value{
A function with signature \verb{(index, levels)} suitable for \code{panel_fun} in \link[ComplexHeatmap:anno_block]{ComplexHeatmap::anno_block}.
}
\description{
Build block fill panel function
}
\examples{
library(ggplot2)
lv <- c("A", "B", "C")
panel_fun <- annotation_block_fill_graphics(
  levels = lv,
  fill_values = c(A = "#1b9e77", B = "#d95f02", C = "#7570b3")
)
ComplexHeatmap::anno_block(panel_fun = panel_fun)
}
