% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{annotation_block_graphics}
\alias{annotation_block_graphics}
\title{Build block panel function}
\usage{
annotation_block_graphics(subplot, name, border = TRUE)
}
\arguments{
\item{subplot}{A single drawable subplot object.}

\item{name}{A name assigned to the generated grob.}

\item{border}{Whether to draw a rectangle border around the block.}
}
\value{
A function with signature \verb{(index, levels)} suitable for \code{panel_fun} in \link[ComplexHeatmap:anno_block]{ComplexHeatmap::anno_block}.
}
\description{
Build block panel function
}
\examples{
library(ggplot2)
p <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
  ggplot2::geom_point()
panel_fun <- annotation_block_graphics(
  subplot = p, name = "demo-block"
)
ComplexHeatmap::anno_block(panel_fun = panel_fun)
}
