% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chinese_colors}
\alias{chinese_colors}
\title{Chinese traditional colors dataset}
\description{
A dataset containing 1058 traditional Chinese colors.
These colors are extracted from those books:
\itemize{
\item{中国传统色-故宫里的色彩美学 (ISBN: 9787521716054)}
\item{中国美色-最中国文化叁 (ISBN: 9781672897198)}
\item{中国颜色 (ISBN: 9787558016479)}
\item{色谱 (Chinese Journal of Chromatography, ISSN 1000-8713)}
\item{中国色彩图典}
}

Thanks to the author of the \href{https://book.douban.com/review/14384195/}{blog} for providing the data.
}
\examples{
cc <- ChineseColors()
widget <- cc$visual_colors(
  title = "中国传统颜色",
  name_type = "chinese"
)
htmltools::browsable(widget)
}
\concept{data}
