% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_legend}
\alias{get_legend}
\title{Extract legend from a plot}
\usage{
get_legend(plot)
}
\arguments{
\item{plot}{A plot object.}
}
\value{
The legend grob.
}
\description{
Extract the legend grob from a plot object.
}
