% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese_colors.R
\name{print.ChineseColors}
\alias{print.ChineseColors}
\title{Print ChineseColors object}
\usage{
\method{print}{ChineseColors}(x, ...)
}
\arguments{
\item{x}{A ChineseColors object.}

\item{...}{Additional arguments (not used).}
}
\value{
Details of the ChineseColors object.
}
\description{
Print ChineseColors object
}
